Refine work to reduce spills use for Store
Adds logic to deal with lvalues with side-effects via a pointer
temporary (based on Steven's advice).  Also refactors the logic for
contructing the runtime call to deal with the additional generality
of the new case (will ease more cases later, too).

This commity also fixes incorrect behavior in the previous code when
the RHS was an lvalue but didn't have the same type as the target.
The fix is a bit too conservative at the moment, as it requires a
spill for the case where the types differ only in signedness.