Create BuildUPCRAdd{Pshared[I1],Shared}() to address issue #86
This commit factors all sites in the code which create calls to any
of the three pointer-to-shared addition routines.  By doing so, we
now can fold patterns of repeated arithmetic as are found in
accesses to multi-dimensional arrays.  Additionally 0-increment
addition is now squashed before it is ever created, allowing removal
of code from FoldUPCRLoadStore() that also squashed 0-increment
adds (but only when they appear directly in a Load or Store).